Changes to HTML and CSS
In 'CustomerOrderListPage.vue' the 'left-div' was removed and the 'middle-div' was renamed 'left-div', changes was also made to the width of the div boxes to make them fit the page better.

In 'KitchenOrder.vue' the orders now have a hover color change and a cursor change to make it easier to see where the cursor is on the page.